Real time implemented

+++ src/policy.hpp
@@ -1,6 +1,9 @@
 #ifndef POLICY_HPP_
 #define POLICY_HPP_
 
+#include <state.hpp>
+#include <action.hpp>
+
 class policy {
 public:
     virtual action apply (const state &s) const = 0;
@@ -8,7 +11,7 @@ public:
     virtual void process_reward(
         const state &s,
         const action &a,
-        double r,
+        unsigned r,
         const state &s_p) = 0;
 };
